+++ lib/capistrano/simple_systemd.rb
@@ -1,4 +1,3 @@
 require 'capistrano/simple_systemd/paths'
 require 'capistrano/simple_systemd/helpers'
-load File.expand_path('../tasks/simple_systemd/global_tasks.rake', __FILE__)
-load File.expand_path('../tasks/simple_systemd/specific_tasks.rake', __FILE__)
+load File.expand_path('../tasks/simple_systemd.rake', __FILE__)