+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # Changelog
+### Unreleased
+- Added the following specific tasks
+  - cap {stage} systemd:{service}:upload to upload a single service file
+  - cap {stage} systemd:{service}:remove to upload a single service file
+- Added the following global tasks
+  - cap {stage} systemd:enable to enable all services
+  - cap {stage} systemd:disable to disable all services
+  - cap {stage} systemd:remove to remove all services
+- Moved both global and specific tasks into one file
+- Pruned dependencies
+
 ### v0.1.1, 2019-10-26 (Maintenance release)
 - Moved gem to new repository (https://github.com/gkosae/capistrano-simple_systemd.git)
 - Added sample service file template to README